Qol Futures to the code. Added a gift system.
New streak counter when a person guesses 2 or more letters in a row.
New gift system that only activates if a person hasn't guessed a letter in three tries. Can choose to accept or decline the gift. By accepting on random they either get one more try or receive a hint letter.
"You should type only a single letter! and "You shouldn't use the same letter again" now appear at the top.
Letters in the word now don't appear in the used list to make it easier to guess.
Made the Output Encoding to UTF-8 to avoid errors.